Update accconfirm.rst
A confirmation email should be resent at this point for cases when the user lost the original email.  Without this email the user would not be able to confirm their identity and they would have no other means to generate an email, they would be stuck in limbo.          

This matches the pattern in Rick Anderson's similar article: http://www.asp.net/mvc/overview/security/create-an-aspnet-mvc-5-web-app-with-email-confirmation-and-password-reset