Added comment asking for clarification of Web root
Apologies in advance if this is not an appropriate way to ask for additional clarification in the documents. I added a comment in square brackets suggesting why I don't understand the explanation of Web root. My intention is that an author who knows the answer would be able to provide more detail and delete the comment.