Added linux production guidance
Add --signal INT for a clean shutdown

"Please add --signal INT to the list of arguments. Right now, start-stop-daemon will send SIGTERM to end the process, but there isn't (usually) any code in ASP.NET that listens to this.
For SIGINT, Hosting adds a CancelKeyPress handler that does the clean shutdown."
@Tragetaschen