Spell check and theme fixes
"upzip" was a bug in an earlier commit in referring to the "unzip" package
Minor fixes after checking the doc out in the sphinx theme to ensure it looked okay.
Removed note that .NET Core dependencies were optional to avoid confusion.  It is in its own section so users can decide to follow the .NET Core path, the Mono path, or both.
Referred to Kestrel the way the Ubuntu section does.