Be more explicit about platforms
The binary versions of DNX that we provide for RC1 and RC1 Update one
have tightly coupled dependencies with versions of the libraries that
ship in Ubuntu 14.04 and CentOS 7. Other distros (or even other versions
of CentOS and Ubuntu) are unlikely to work out of the box.

Update the wording to be a little more clear about this.