Add Nicholas Blumhardt DI blog
From: David Fowler Sent: Thursday, March 10,
Subject: Excellent article on dependency injection

As we're making DI a first class citizen in the new ASP.NET stack, it's important to understand some of the reasoning and patterns behind this. I just found this article written by autofac author and ex msft employee Nicholas Blumhardt. It's a great way to get a proper mental model of a dependency injection system and container:

http://blogs.msdn.com/b/nblumhardt/archive/2008/12/27/container-managed-application-design-prelude-where-does-the-container-belong.aspx