Update static-files.rst
Hopefully these changes will help other developers implement with a bit less friction.
The title 'default files' is fine, but most experienced ASP.NET developers are scanning for the word "default document".
I Also, added a note about the order in which default and static is called.  It makes a difference see the SO link below.  
http://stackoverflow.com/questions/30228620/kestrel-on-aspnet-vnext-doesnt-serve-index-page-under