Mention of a small caveat for static file serving
When reading this document and trying to build the app from scratch users could forget or not know that root content directory needs to be added to the WebHostBuilder. It should be mentioned somewhere in this document. On a Mac for example there is no VisualStudio and no template projects that have the correct initial setup.