Added note RE: @RenderSection scripts
Found the solution from this StackOverflow post: http://stackoverflow.com/questions/10971149/the-following-sections-have-been-defined-but-have-not-been-rendered-for-the-layo

It means that you have defined a section in your master Layout.cshtml, but you have not included anything for that section in your View.

The final div within _Layout.cshtml should look like this:

    <div class="container body-content">
        @RenderBody()
        @RenderSection("scripts", required: false)
        <hr />
        <footer>
            <p>&copy; 2016 - MvcMovie</p>
        </footer>
    </div>

Another user reported the same issue within the comments on this post. I replicated the issue myself, and this resolved it.