Added info about installing .NET Core SDK for Windows
The instructions don't specify that you need to also have the .NET Core SDK for Windows installed. Making sure that it's explicit from the documentation that that is a prerequisite on the machine.

Added troubleshooting information for the _0x80072efd_ error code, which occurs when you don't have the .NET Core SDK installed prior to installing the Hosting Bundle.