Hopefully reverting earlier change wanting "It is important to remember..."
Add reminder to user to include current assembly in _ViewsImport.cshtml

It does not seem at all discoverable that to add your own tag helpers you need to add a line to _ViewImports.cshtml (I struggled with this for a while before figuring it out).  I believe the doc is currently not sufficient to address this so I expanded the explanation with a likely failure mode the reader my fall in to (as I did). (reverted from commit 8ecf93b13a1966c9e3b886ed22ff899c1d103100)