Fix explanation: setting a default selected option
The current doc treats the second parameter of @Html.DropDownList() as a way to set the default selected option. It gives an example of choosing "Comedy" by default. In fact, the given example creates a second entry that says "Comedy" but has an empty value. I provided a correct example.