Update Localization sample app to 2.0 (#4219)
* Update Localization sample app

Fixes #2943

* Takes the sample to 2.0
* HTML5 (I gave this some thought, and I still prefer unclosed void
elements. Content-Type = "text/html" and doctype is HTML5. I think the
spec (making the closure optional) is a bit of a hack on their part.
These tags do not contain content, so imo they really shouldn't be
closed (logically).
* Bower out: Adds too many complications. We have *www/lib* locked out
of the source; therefore, the sample will break if the machine is
offline and Bower isn't run. Bower requires npm. Many *many* devs and
shops don't like it and don't want it. It has nothing to do with
localization. imo samples should run everywhere possible with minimal
config and setup and only use the tech required to demo the focused
tech.
*

Add one file

Updates

* React to feedback

Also patched the title, description, and keywords on this pass.