Update apache-proxy.md (#4291)
The example systemd unit file causes an error if copied directly and used as is (which a lot of first time users likely will).

Although the example service will run it will cause a systemd error because comments in systemd unit files need to be on a newline. In this case I've moved the comment to the line above. Now when you do `systemctl status kestrel-hellomvc.service` you won't see an error in the output.