Suggestion for better wording.
While going through the tutorial on Mac, I was confused by this wording. I assumed that "command window" was something that was part of Visual Studio.

It may be clearer (especially to newcomers) to be consistent with the wording that's only a few lines above "From the command line, run the following .NET Core CLI commands:".

This is just a suggestion, now I realize that the words "Open a command window" appear in many parts of the documentation.