Fix link to writing style guide in CONTRIBUTING.md
The link to Microsoft Writing Style Guide was pointing to `/style-guide/welcome/` path that does not exist in this project's repository. As I have not found the GitHub project that contains the source for Microsoft Writing Style Guide, I modify the link to point to the published version at Microsoft website.