Update live with current master (#5898)
* Update azure-apps-preview-notice.md (#5857)

* Link to .NET Core CLI publish command doc (#5864)

* Link to .NET Core CLI publish command doc

* Verbiage tweak

* fix typo (#5866)

* Update intro.md (#5869)

`classnameID` (where `classname` is the name of the class, such as `Student`) as the primary key. In `classnameID`, `classname` is the name of the class, such as `Student` in

* Update view-components.md (#5872)

altering range of code displayed to remove hanging closing } and closing table tag

* C13511:  Spaces should be replaced to avoid localization problems (#5871)

Hello, @rick-anderson, 

Localization team has reported source content issue that causes localized version to have broken format compared to en-us version.  

Please, help to check my proposed file change into the article and help to merge if you agree with fix. If not, please, let me know either if you would like me to fix it in another way within this PR, if you prefer to fix it in another PR or if I should close this PR as by-design. In case of using another PR, please, let me know of your PR number, so we can confirm and close this PR. 

Many thanks in advance.

* CI Update (#5878)

* Recommended language change (#5883)

* Updated with NET Core SDK 2.0  commands (#5886)

* Updated with NET Core SDK 2.0  commands

* Quick UE pass

* Prevent javascript injection. (#5885)

* Troubleshoot SDK warnings (#5880)

* WIP Troubleshoot SDK warnings

* Work

* work

* Address Gauntlet warning in ms.date value

* Adjust H2 title casing and spacing

* Quick UE pass

* Update Configuration doc to use double underscores instead of colons for hierarchical config keys (4968) (#5876)

* changes to configuration doc

* fixes for scott

* change for Dan

* Update index.md

* Code fence double underscore and add missing period

* Eliminate unnecessary sentence

* Missing npm task runner extension (#5893)

* Revert "Missing npm task runner extension (#5893)" (#5894)

This reverts commit a90faebff2fa1e2861d38e56b3fa86fb6294c49d.

* Note about task runner explorer option (#5895)

* Note about task runner explorer option

The gulp file has to be in the root directory or in the directory with package.json not in any subdirectory.

* verbiage tweaks

* added anchored link (#5896)