Fixed hidden element rendered in md
The span is invisible on the page. Also, everything in the intro recommending using bower "Adding Bootstrap to an ASP.NET Core project is simply a matter of adding it to bower.json as a dependency:" should be updated to nuget and csproj. Also the ASP.NET CORE WEB application template no longer contains a login on VS 15.7 preview 3 so maybe just delete the reference?