fixed globbing pattern.
Good information. The old globbing pattern (ignoring the right parenthesis typo) would match every input so I fixed it. A third way I thought to do it would be: `/**/!(*.!(css))` which prevents min because the *. before the parenthesis will match only the first . and not the last `.`. Negating twice gets the .css extension back. I decided for the other way because this will also match .css in the middle of words.