Update create-the-project.md (#6120)
* Update create-the-project.md

Four months ago, in January, someone asked where the Models folder is at and the answer was to change the authentication. I assume the instructions should have that information and apparently it has been missing since the beginning. For VS 2017 I think the "Create Project" button is now an "OK" button but I don't know if my update about that is relevant.

* Verbiage tweaks