Explain how to get bootstrap.cs for local use
Downloading the github sample doesn’t include the bootstrap CSS. The layout CSHTML loads a local copy of the bootstrap CSS file when in “Development”, so I think the docs should be updated to mention this. I had to add the CSS file manually in my project to resolve this.