Update anti-request-forgery.md
Clarify that there are alternative scenarios a malicious site could implement, in addition to the example of a basic CSRF attack. The original content conveys the impression that the user is required to interact with the malicious site, which is presumably almost never the case when the user has JavaScript enabled.