C26450: modifying URL format
Hello, @rick-anderson ,
Localization team has reported source content issue that causes localized version to have broken/different format compared to en-us version.
"URL is not well formatting if this is intended to be shown as a link"

Please review and merge the proposed file change to fix to target versions. If you make related fix in another PR then share your PR number so we can confirm and close this PR.
Many thanks in advance.