Logging topic enhancements (#7386)
Fixes #7371 

[Internal Review Topic](https://review.docs.microsoft.com/en-us/aspnet/core/fundamentals/logging/index?view=aspnetcore-1.0&branch=pr-en-us-7386)

Will be doing more to this topic when it comes up for the sample update. For now, surface more info on the appsettings config, update info on `IncludeScopes`, and drop the tabs.

For engineering review, the main update is the section that starts at Line 74.

Thanks @totht91 for surfacing the lack of coverage here. 🎸