Update webapp-alias-notice.md (#7882)
* Update webapp-alias-notice.md

The original note specified "2.1", which wasn't specific enough (I had 2.1.200 and wondered why it wasn't working). This generalizes things, moves the most important part to the front and changes the download link to the .NET Core download home page instead of the version list.

If you want to keep the link pointed at the version list, I suggest https://www.microsoft.com/net/download/dotnet-core/2.1. The current link points directly to a specific version that is already out of date.

* Verbiage tweaks