Added note for working with virtual directories
I spent a day or more researching this so that I could use Swagger-UI in my ASP.NET Core API which is used with a reverse-proxy and virtual directory.  In the end found Swagger-UI requires relative URLs to start with a period.  Related to https://github.com/swagger-api/swagger-ui/issues/4107

Hoping this documentation note will help others.