Update ws-federation.md
This documentation page appears to contain a minor mistake: it claims that the sample code provided for integrating WS-Federation with ASP.NET Core Identity should be added to the `Configure` method, not `ConfigureServices`. In addition to the `Configure` method not normally an IServiceCollection object handy, this seems to disagree with both the code sample below on this page for the example without Core Identity and with every other example of using WS-Federation with Core Identity I can find online.