Remove Angular SSR documentation (#11151)
Fixes #10501 and #8139.

This PR removes the Angular SSR documentation from the docs. When the Angular template initially shipped it included SSR as a built-in feature. Since we've moved the Angular template to be based on the Angular CLI, SSR is no longer included. Instead of trying to maintain our own instructions in the ASP.NET Core docs for enabling Angular SSR we recommend referring to the [Angular docs directly](https://angular.io/guide/universal).