Update get-started-with-the-ajax-control-toolkit-cs.md
I assume my wording needs improving but I think it helps to know how easy it is to do.

For some reason the toolkit was not added to my toolbox the first time so I downloaded (installed) again and then it got added. So perhaps I did something (that these instructions say to do) that I should not have and if so then it is especially important to say exactly what we are supposed to do.