Update index.md
Some Microsoft people are confused about what documentation is and what tutorials are. Perhaps I should not say that that way but I hope it is clear what the differences are. It is understood that tutorials do not provide all the details and documentation is expected to provide all the details and can therefore be confusing sometimes. Tutorials are written to omit details to aid in understanding. Tutorials should not be considered documentation.