April SEO Effort - Changed description metadata
April SEO sweep: High bounce.  The already helpful topic description we would like to show up in the search results is getting bypassed in Google and instead it is showing some random info halfway down the page.  Updating the description meta data, date and intro with a minor change to see if we can get the correct description to show up.  For all April SEO sweep changes we will follow up in month to verify it is going in the right direction.