Fix .NET Standard issue and link version table
You can't build "App" with .NET Standard, as it is not a runtime. You can only build libraries with it. And those libraries can be executed by a runtime that implements .NET Standard, like .NET Core or .NET Framework. So I changed "Apps" to "Libraries" in that paragraph. I also linked the versions table from the .NET Standard GitHub repo. 

This change is a minor fix to the already closed issue #11866