Workaround designer issue in VS (#12823)
Because it inherits from `ServiceBase` (via `WebHostService`), VS treats this `CustomWebHostService` class like a "Component" and tries to open a design surface for it. This causes problems because the designer can't load properly over this type (it has a constructor). We are fixing `WebHostService` in 3.0 (and it's no longer recommended anyway, with Worker Service), but the 2.x sample should be updated to lead people down the happy path.