Update health-checks.md
This proposal attempts to simplify the sample code in the `ResponseWriter` delegate to be a little easier to read and remove the numerous new JProperty and JObject invocations. The `StringEnumConverter` is passed into the serializer to ensure the `HealthStatus` enum's are rendered by their names and not their values (Healthy or Unhealthy vs. 2 or 0).
Thank you.