Change how the tool is launched
With https://github.com/dotnet/HttpRepl/pull/223, we changed the name of the tool from `dotnet-httprepl` to `httprepl`. This changes the way the tool is executed from `dotnet httprepl` (or `dotnet-httprepl`, though that wasn't used in the docs) to just `httprepl`.
This PR updates the docs to reflect that change.