Patch 443 default line
Addresses #15249

This is a quick patch to correct an inaccuracy wrt default being 443. This sentence isn't necessary. Let's 🔪 it. The *Port configuration* section earlier is where the logic of port selection is handled.

I'll also paste the source cross-links into #15249 so that it will be fast 🏃 to get the latest framework logic for that issue.