Temporary no-loc Razor term fix for high vis issue
Changing the title meta data so that "Razor" is lower case as "razor" as recommended by the localization team.  There is currently an issue where no-loc: topic meta data terms are ignored if the same word in the same case exists in the title meta data.   

For this particular topic's case, it ends up with the translation of Razor into "Shaving" and similar translations.  The support team and product unit would like it fixed quickly and the community has complained about it.

Localization is working on a longer term fix for this issue.