Apply suggestions from code review
Thanks @scottaddie. btw - Are u suggesting with your non-slashed suggestions that you don't care if the non-slashed URL generates an extra 301 redirect? I always (try to) supply links with the fewest roundtrips, so I'm not clear on what seems to be a blanket rule that you have based on prior reviews/discussion.

Co-authored-by: Scott Addie <10702007+scottaddie@users.noreply.github.com>