Add JavaScript extension message
Was working with a customer today who has v1.45 installed and we didn't see the "debug preview" option in settings.  We didn't know we had to upgrade to v1.47 and based on what was in the docs we were at a dead end.  Suggest telling user that either v1.47 or the extension is required.

azsdke2e