Update additional-scenarios.md
Added a using directive to the "Attach tokens to outgoing requests" sections in the first sample. 
I have noticed that it is not possible to call ```builder.Services.AddHttpClient()``` without the additional using.

It is not very hard to figure out the missing packages name but maybe referring it in the documentation, makes it easier for beginners to get started.