Update hosted-with-identity-server.md
The code identity.RemoveClaim(existingClaim); was throwing an exception that the collection was being altered. When I commented that line of code out everything was fine but claims were not being removed. The modification above creates a separate list so as not to involve the Claims collection on User.Identity.