Error in Section "Test the PostTodoItem method 2.1"
When following along with the .NET Core 2.1 tutorial, the part that says to "Set the URI to https://localhost:<port>/api/TodoItem..." fails. However, if you look in the screenshot for that part, the URI does not say "TodoItem" but is instead "Todo". When you use the URI "https://localhost:<port>/api/Todo", the test passes.

I had opened issue #20540 first about this change before I saw that I could just propose the edit here.