Fix Readme.MD, add correct startup file for CORS
The sample app contains two files that contain a `Configure` method:
- `Startup.cs`
- `StartupJavaScript.cs`

Hitherto, the Readme said:
>Add the following CORS middleware configuration to the web API's `Startup.Configure` method:</p>

But that is incorrect. The configuration needs to be added to `StartupJavaScript.Configure`. It took me ages to figure it out, trusting the Readme to be correct - when it wasn't.

There's no harm in making this change. Yes, people can eventually try it out themselves, but there's no harm in adding the correct file reference.