Add information about the location of app.UseWebsockets()
See https://github.com/dotnet/aspnetcore/issues/30646 - you won't be able to accept WebSocket connections in a controller if the UseWebSockets() middleware is below the UseEndpoints() middleware. In my opinion, whether or not the root issue is a bug is irrelevant (and I don't know if it is or not yet). This little note would benefit documentation readers. :)