Update name used in cookie handling snippet
'UseCookies = false' should be the default from a security/privacy perspective, but due to backwards compatibility probably can't be changed now. Thus, it's important that the example here show how to change this for the default instance since that is the most common use case and it was difficult/took some time to find how to configure this for the default instance (since you can't call 'ConfigurePrimaryHttpMessageHandler' on `services.AddHttpClient()`. By updating this example, it helps show you need to use `Options.DefaultName` to make this change for the unnamed/default instance. Related to #15848.