Update directory-structure.md (#21341)
* Update directory-structure.md

For self-contained deployments (at least on ASP.NET Core 5.0) the *.exe extension is only generated for Windows. For Linux (and presumably macOS) no extension is generated. This behaviour is consistent with that for framework-dependent deployments so I just copied the relevant text. I am not sure if the behaviour applies to versions < aspnetcore-3.0 so I haven't touched the text there.

It may also be useful to add that if you choose the `Produce single file` option, none of the `*.dll` files remain in the publish directory (I love this feature!).

* Update directory-structure.md

Co-authored-by: Rick Anderson <3605364+Rick-Anderson@users.noreply.github.com>