Update hosted-with-azure-active-directory.md
The document, in the client configuration section reference the auto generated code in the program.cs. In this code it specify the template will produce a name based on the App Assembly concatenated with the ".WebAPI" literal. This has changed in recent versions and the reference should be App Assembly concatenated with the ".ServerAPI" literal.