Use ISO date syntax in web-config.md
The section on `stdoutLogFile` shows how a log file saved on `2/5/2018` would be named `stdout_20180205194132_1934.log`. For international readers like me, its not clear whether `2/5/2018` means 2/May/2018 or 5/Feb/2018. To remove the ambiguity, I have tweaked the example to use the ISO date `2018-02-05` which is how it appears in the log file name as well.