Fix broken internal document links
These links are broken in the [microsoft.com documentation page](https://docs.microsoft.com/en-us/aspnet/core/blazor/javascript-interoperability/?view=aspnetcore-5.0). Inspecting the header links in the document reveals that some of them have the characters `-1` appended, which keeps the links in the provided index from working.
I have added `-1` to the links as they appear in the documentation page.