Improve launch.json examples
Previously, a project created by "dotnet new" and opened in VS Code was unable to debug.  Specifically, when the debugger was started, a new browser window would open "about:blank, wait for 30 seconds, and then close.  The issue was that:
- The launchSettings.json generated by the "dotnet new" command uses a random port number.
- The launch.json file generated by VS Code omits the "url" parameter (the default value is "https://localhost:5001").
The fix is to add the "url" parameter to launch.json.

Also, I merged the "Example launch configurations" section with the relevant steps above.  This makes it easier for users to find the appropriate config, as well as simplifying long-term maintenance of the documentation.