Add warning about limited range of decimal encoding
I was pleasantly surprised to see that Microsoft provides guidance on how to encode decimal in protobuf but I quickly realized that this encoding only supports a limited range of values with a limited precision. While this limit probably is fine in most cases I think it should be mentioned so I added a warning and placed it after the note that hints at the superiority of this encoding to contextualize it a bit and also help the "copy-paste developer" to better understand the encoding at a glance.