Add missing host & fixed number of hosts available (#25922)
* Add missing host & fixed number of hosts available

The original section said: 
There are three different hosts:

.NET WebApplication Host
.NET Generic Host
ASP.NET Core Web Host
The .NET Minimal Host is recommended and used in all the ASP.NET Core templates. 
----

Notice a 4th host, ".NET Minimal Host" is discussed under the list. I added it to the bulleted list.

The strart setence said "There are three different hosts:" but the number three is wrong (the correct number is four) but I made the setence agnostic to the number of hosts in case a 5ht, 6th or 7th host is added. The new setence is "The different hosts include:" which works for 50 or 100 or 1000 hosts.

* Update index.md

* Update aspnetcore/fundamentals/index.md

* Update aspnetcore/fundamentals/index.md

Co-authored-by: Rick Anderson <3605364+Rick-Anderson@users.noreply.github.com>