No need to explicitly call RemoveFromGroupAsync
The sample added this explicit `RemoveFromGroupAsync` call when the connection disconnects, however, it is a redundant call since when the connection disconnects, it is automatically removed from the group. 
We have customers using Azure SignalR followed this sample and invokes `RemoveFromGroupAsync` for every client disconnect, this increases unnecessary burdon to the Azure SignalR service when the client connection closes burst.