No need to explicitly call RemoveFromGroupAsync (#26394)
* No need to explicitly call RemoveFromGroupAsync

The sample added this explicit `RemoveFromGroupAsync` call when the connection disconnects, however, it is a redundant call since when the connection disconnects, it is automatically removed from the group. 
We have customers using Azure SignalR followed this sample and invokes `RemoveFromGroupAsync` for every client disconnect, this increases unnecessary burdon to the Azure SignalR service when the client connection closes burst.

* Update hubs.md

Section Handle events for a connection for all versions: Noted no need for RemoveFromGroupAsync in OnConnectedAsync

* Update hubs.md

Added note on RemoveFromGroupsAsync to all versions of topic.  Corrected Typo.

* Update aspnetcore/signalr/hubs.md

Improved statement.

Co-authored-by: Brennan <brecon@microsoft.com>

* Update hubs.md

Updated statement on RemoveFromGroupAsync for all topic versions.

* Update gRPC example for StronglyTypedChatHubs.cs

Removed RemoveFromGroupAsync call from OnDisconnectedAsync.

Co-authored-by: Wade Pickett <wpickett@microsoft.com>
Co-authored-by: Brennan <brecon@microsoft.com>