update custom provider code snippet
The code on line 355 lacks an awaiter despite being part of an async method, and triggers a warning for not being awaited in .NET 6. I settled on this implementation which eliminates the warning and functions properly. It seems like there should be some more code here to help with the async method, hence my proposed change. However, please let me know if I am incorrect or misunderstanding this code.