Rearrange text for clarity (#27778)
* Rearrange text for clarity

This article is excellently written. 
It is important to bear in mind that auth is one of the most challenging concepts for developers, particularly with the multitude of technologies and terms. I have read this article multiple times and still get stuck on this line, wondering whether Microsoft Identity Platform is the same as ASP.NET Core Identity.
I would suggest that the proposed change makes this much clearer.

* Updates

Co-authored-by: Luke Latham <1622880+guardrex@users.noreply.github.com>