Update virtualization.md to inform about the need to set the renderModel to InteractiveServer
I have come to the fact that was trying to setup a project with .NET 8 to use virtualize and I only had the configuration on program.cs to AddInteractiveServerComponents.  Only with that configuration Virtualize was not rendering the elements. For sake of clarity I belive that is important to inform developers that Virtualize only works if the rendermode is interactiveServer.