Make alternatives 3b (non-DI) and 3c (DI) functionally identical
Currently, documentation reads as if in order to use DI in a custom `IOutputCachePolicy` consumers must register the policy via `AddBasePolicy(...)`, which is incorrect. This is misleading and may lead consumers to over-cache all endpoints and actions when they only want to override the default policy, while keeping the same behaviour as the default policy which requires the `[OutputCache]` attribute and gives more granular caching control.

The way to apply a policy on all of app's endpoints is already documented separately above: https://learn.microsoft.com/en-us/aspnet/core/performance/caching/output#default-output-caching-policy.