Reorganize ASP.NET Framework to Core migration guidance
This change reworks the ASP.NET Framework to Core migration documentation to tell the story of how to migrate better. The goal is to bring together the different approaches in a way that will quickly let developers understand what their options are and how to make the best decisions for their migration journey.