Fix SetDefaultPolicy being used like SetFallbackPolicy
The description mistakenly says `SetDefaultPolicy` twice, despite describing the behavior of `SetFallbackPolicy` in the first sentence and then the behaviour of `SetDefaultPolicy` in the next sentence. I suppose this is a typo.

What it used to say is false, as changing `SetDefaultPolicy` will never impact endpoints without an `[Authorize]` attribute.

This is especially dangerous, as someone copying the example code may just naively assume their API is already secure without using `[Authorize]`, while it obviously isn't.

Additionally, changed the example below to show how to require authenticated user by default, similar to how it's presented in the linked require-authenticated-users article section.

This was done as the previous example using `SetDefaultFallback` essentially did nothing, as requiring authenticated users is what the default policy does already with nothing set, so the old example is of no use.